refactor(tags): centralize tag addition logic in infrastructure and pipeline stacks
Refactored the code for adding tags to AWS resources by creating a private method addTags(). This change affects both the BlogInfrastructureStack and BlogPipelineStack classes. The previous inline for-loop that iterated over the props.tags object has been removed and replaced with a call to this.addTags(props?.tags), reducing duplication and improving maintainability. The private addTags() method encapsulates the functionality required to add tags, making the codebase cleaner and easier to manage.