feat(apigateway): extend referer policy with non-www domain
Added 'playingaws.com' and 'playingaws.com/*' to the allowed referers for the API Gateway resource policy to ensure requests originating from the root domain are permitted access, in addition to those from the 'www' subdomain. This change allows users to access the API regardless of whether they include 'www' in the URL or not.